+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Tests/Mocks/Services/MockDevOpsService.cs
@@ -13,6 +13,11 @@ namespace Azure.Sdk.Tools.Cli.Tests.Mocks.Services
             throw new NotImplementedException();
         }
         
+        Task<List<ReleasePlanDetails>> IDevOpsService.ListOverdueReleasePlansAsync()
+        {
+            return Task.FromResult(new List<ReleasePlanDetails>());
+        }
+        
         public Task<PackageWorkitemResponse> GetPackageWorkItemAsync(string packageName, string language, string packageVersion = "")
         {
             throw new NotImplementedException();