+++ src/dotnet/APIView/APIViewIntegrationTests/RepositoryTests/CosmosReviewRepositoryTests.cs
@@ -28,7 +28,8 @@ namespace APIViewIntegrationTests.RepositoryTests
             _cosmosDBname = "CosmosReviewRepositoryTestsDB";
             config["CosmosDBName"] = _cosmosDBname;
 
-            _cosmosClient = new CosmosClient(config["CosmosEndpoint"], new DefaultAzureCredential());
+            var credential = new ChainedTokenCredential(new AzureCliCredential(), new AzureDeveloperCliCredential());
+            _cosmosClient = new CosmosClient(config["CosmosEndpoint"], credential);
             var dataBaseResponse = _cosmosClient.CreateDatabaseIfNotExistsAsync(config["CosmosDBName"]).Result;
             dataBaseResponse.Database.CreateContainerIfNotExistsAsync("Reviews", "/id").Wait();