+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.10 (Unreleased)
 
 ### Features Added
+- Add CLI command to identify in progress release plans with past due date
 
 - Allow non-exact matches for package name in `azsdk_release_sdk` tool
 - Moved `azsdk_check_api_spec_ready_for_sdk` and `azsdk_link_sdk_pull_request_to_release_plan` under release plan command hierarchy.