+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 - Add CLI command to identify in progress release plans with past due date
 
+- Allow non-exact matches for package name in `azsdk_release_sdk` tool
 - Moved `azsdk_check_api_spec_ready_for_sdk` and `azsdk_link_sdk_pull_request_to_release_plan` under release plan command hierarchy.
 - Added APIView tools to expose APIView functionality to MCP agents (`get-comments`) and CLI (`get-content`)