+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -8,7 +8,7 @@
 - Allow non-exact matches for package name in `azsdk_release_sdk` tool
 - Moved `azsdk_check_api_spec_ready_for_sdk` and `azsdk_link_sdk_pull_request_to_release_plan` under release plan command hierarchy.
 - Added APIView tools to expose APIView functionality to MCP agents (`get-comments`) and CLI (`get-content`)
-- Update `azsdk_verify_setup` to check that `core.longpaths` in git config is set to true on Windows
+- Updated `azsdk_verify_setup` to check that `core.longpaths` in git config is set to true on Windows
 
 ### Breaking Changes