+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 0.5.11 (unreleased)
+
+### Features Added
+
+- Updated `azsdk_verify_setup` to check that `core.longpaths` in git config is set to true on Windows
+
 ## 0.5.10 (2025-12-08)
 
 ### Features Added
@@ -8,7 +14,6 @@
 - Allow non-exact matches for package name in `azsdk_release_sdk` tool
 - Moved `azsdk_check_api_spec_ready_for_sdk` and `azsdk_link_sdk_pull_request_to_release_plan` under release plan command hierarchy.
 - Added APIView tools to expose APIView functionality to MCP agents (`get-comments`) and CLI (`get-content`)
-- Updated `azsdk_verify_setup` to check that `core.longpaths` in git config is set to true on Windows
 
 ### Bugs Fixed